# 2015-2016-api-mobiliteitscentrum/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mobiliteitscentrum Gent Website API (VOPRO 2015-1016)",
        "description": "API specificatie voor Milestone 3.\n",
        "version": "3.0.0"
    },
    "host": "voproX.ugent.be",
    "schemes": [
        "http"
    ],
    "basePath": "/api",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/user/{user_id}/travel/": {
            "get": {
                "summary": "Returns all travels on which this user is subscribed.",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All travels of this user",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Travel"
                            }
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Creates a new travel for this user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "travel",
                        "in": "body",
                        "description": "The new travel's parameters",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Travel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The created travel",
                        "schema": {
                            "$ref": "#/definitions/Travel"
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User or embedded route does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "The travel already exists  (all properties already exist in another travel)\n",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/user/{user_id}/travel/{travel_id}/route/{route_id}/": {
            "get": {
                "summary": "Returns the specific route on this travel",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "travel_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "route_id",
                        "in": "path",
                        "description": "The unique identifier of this route",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The route",
                        "schema": {
                            "$ref": "#/definitions/Route"
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User, Travel or Route does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Edits a specific route of this travel",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "travel_id",
                        "in": "path",
                        "description": "The unique identifier of this travel",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "route_id",
                        "in": "path",
                        "description": "The unique identifier of this route",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "route",
                        "in": "body",
                        "description": "Route input parameters",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Route"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The modified route",
                        "schema": {
                            "$ref": "#/definitions/Route"
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User, travel or route does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "The route already exists",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deletes the route from the travel",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "travel_id",
                        "in": "path",
                        "description": "The unique identifier of this travel.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "route_id",
                        "in": "path",
                        "description": "The unique identifier of this route",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The route was deleted"
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User, Travel or Route does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Address": {
            "type": "object",
            "properties": {
                "street": {
                    "type": "string"
                },
                "housenumber": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string",
                    "description": "2 letter country code (ex. BE)",
                    "maxLength": 2,
                    "minLength": 2
                },
                "postal_code": {
                    "type": "string"
                },
                "coordinates": {
                    "$ref": "#/definitions/Coordinates"
                }
            },
            "required": [
                "street",
                "housenumber",
                "city",
                "country",
                "postal_code"
            ]
        },
        "Coordinates": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number",
                    "format": "double",
                    "description": "Latitude"
                },
                "lon": {
                    "type": "number",
                    "format": "double",
                    "description": "Longitude"
                }
            },
            "required": [
                "lat",
                "lon"
            ]
        },
...
        "Route": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Unique identifier for  this route"
                },
                "waypoints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Coordinates"
                    }
                },
                "transportation_type": {
                    "type": "string",
                    "description": "Indicates which type of transportion this route uses (ex. bus, car, bike)"
                },
                "notify_for_event_types": {
                    "type": "array",
                    "description": "User has to mark specific EventTypes as relevant  for this route to receive notifications for events of these types",
                    "items": {
                        "$ref": "#/definitions/EventType"
                    }
                },
                "notify": {
                    "type": "object",
                    "description": "The mediums on which the user wishes to be notified of events on this route",
                    "properties": {
                        "email": {
                            "type": "boolean",
                            "description": "User wishes to be notified via email"
                        },
                        "cell_number": {
                            "type": "boolean",
                            "description": "User wishes to be notified via SMS"
                        }
                    }
                },
                "active": {
                    "type": "boolean",
                    "description": "Boolean to indicate if the user wants to receive notifications for this route"
                }
            },
            "required": [
                "waypoints",
                "transportation_type",
                "notify_for_event_types",
                "notify",
                "active"
            ]
        },
...
        "EventType": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "description": "Type of event"
                }
            }
        },
        "Travel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Unique identifier for this Travel"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Link"
                    }
                },
                "name": {
                    "type": "string",
                    "description": "The user can enter a name to recognize this route."
                },
                "time_interval": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "format": "hh:mm"
                    },
                    "description": "2 elements in 24 hour notation (ex. 08:30)"
                },
                "is_arrival_time": {
                    "type": "boolean",
                    "description": "Indicates if the above time is for arrival or departure"
                },
                "recurring": {
                    "type": "array",
                    "description": "The 7 days of the week. If the element is true, this travel is repeated each week on this day",
                    "items": {
                        "type": "boolean"
                    }
                },
                "startpoint": {
                    "$ref": "#/definitions/Address"
                },
                "endpoint": {
                    "$ref": "#/definitions/Address"
                }
            },
            "required": [
                "name",
                "startpoint",
                "endpoint",
                "time_interval",
                "is_arrival_time",
                "recurring"
            ]
        },
        "Link": {
            "type": "object",
            "properties": {
                "rel": {
                    "type": "string",
                    "description": "Indicates the relation of the linked object (ex. self)"
                },
                "href": {
                    "type": "string",
                    "description": "The actual link in full URL (ex. https://vopro.ugent.be/api/user/123/)"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32",
                    "description": "HTTP status code (ex. 400)"
                },
                "message": {
                    "type": "string",
                    "description": "Human readable error (ex. You are not logged in)"
                },
                "fields": {
                    "type": "string",
                    "description": "Missing or incorrect fields for input"
                }
            }
        },
...
    }
}